AI/requirements.txt
python-chess
stockfish